// src/lib/types/http-layer.d.ts
import IHttpHandler from './http-handler'
import IHttpRequest from './http-request'
import IHttpResponse from './http-response'
import INext from './next'
import IHttpRoute from './http-route'

export default interface IHttpLayer {

  route: IHttpRoute

  parsePathParameters(path: string): { [name: string]: string }

  match(path: string): boolean

  handle(req: IHttpRequest, res: IHttpResponse, next: INext, error?: Error): void

  isErrorHandler(): boolean

}